+++ src/rouge/core/workflow/steps/__init__.py
@@ -6,13 +6,13 @@ the WorkflowStep interface for the pluggable pipeline architecture.
 
 from rouge.core.workflow.steps.acceptance import AcceptanceStep
 from rouge.core.workflow.steps.classify import ClassifyStep
+from rouge.core.workflow.steps.code_quality_step import CodeQualityStep
 from rouge.core.workflow.steps.code_review import CodeReviewStep
-from rouge.core.workflow.steps.fetch import FetchIssueStep
+from rouge.core.workflow.steps.fetch_issue_step import FetchIssueStep
 from rouge.core.workflow.steps.fetch_patch import FetchPatchStep
 from rouge.core.workflow.steps.implement import ImplementStep
 from rouge.core.workflow.steps.plan import PlanStep
 from rouge.core.workflow.steps.pr import PreparePullRequestStep
-from rouge.core.workflow.steps.quality import CodeQualityStep
 from rouge.core.workflow.steps.review_fix import ReviewFixStep
 from rouge.core.workflow.steps.setup import SetupStep
